fix(genesis): refactor genesis contracts and examples
