--- contracts/kzg-point-evaluation/build.rs
@@ -1,3 +1,0 @@
-fn main() {
-    fluentbase_build::build_wasm_program_from_env();
-}
