+++ contracts/kzg/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "fluentbase-contracts-kzg"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+fluentbase-sdk = { workspace = true }
+revm-precompile = { workspace = true, features = ["kzg-rs"] }
+
+[dev-dependencies]
+
+[build-dependencies]
+fluentbase-build = { workspace = true }
+
+[lib]
+crate-type = ["cdylib"]
+path = "lib.rs"
+
+[features]
+default = ["std"]
+std = [
+    "fluentbase-sdk/std",
+    "revm-precompile/std",
+]
