+++ contracts/kzg/Makefile
@@ -1,0 +1,3 @@
+.PHONY: lib.wasm
+lib.wasm:
+	cargo build
