+++ contracts/kzg/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    fluentbase_build::build_wasm_program_from_env();
+}
