+++ e2e/Cargo.toml
@@ -27,7 +27,6 @@ fluentbase-examples-erc20 = { path = "../examples/erc20" }
 fluentbase-examples-panic = { path = "../examples/panic" }
 fluentbase-examples-greeting = { path = "../examples/greeting" }
 fluentbase-examples-keccak256 = { path = "../examples/keccak256" }
-fluentbase-examples-echo = { path = "../examples/identity" }
 fluentbase-examples-router-solidity = { path = "../examples/router-solidity" }
 fluentbase-examples-rwasm = { path = "../examples/rwasm" }
 fluentbase-examples-simple-storage = { path = "../examples/simple-storage" }
@@ -37,7 +36,7 @@ fluentbase-examples-checkmate = { path = "../examples/checkmate" }
 fluentbase-examples-client-solidity = { path = "../examples/client-solidity" }
 fluentbase-examples-storage = { path = "../examples/storage" }
 fluentbase-examples-json = { path = "../examples/json" }
-fluentbase-examples-nitro-verifier = { path = "../examples/nitro-verifier" }
+fluentbase-contracts-nitro = { path = "../contracts/nitro" }
 
 [features]
 default = ["std", "fluentbase-runtime/debug-print"]
