--- examples/identity/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "fluentbase-examples-echo"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-fluentbase-sdk = { workspace = true }
-
-[build-dependencies]
-fluentbase-build = { workspace = true }
-
-[dev-dependencies]
-hex-literal = "0.4.1"
-hex = "0.4.3"
-
-[lib]
-crate-type = ["cdylib"]
-path = "lib.rs"
-
-[features]
-default = ["std"]
-std = [
-    "fluentbase-sdk/std"
-]
