--- examples/nitro-verifier/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "fluentbase-examples-nitro-verifier"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-fluentbase-sdk = { workspace = true }
-hex-literal = { version = "0.4.1", default-features = false }
-coset = { version = "0.3.8", default-features = false }
-ciborium = { version = "^0.2.1", default-features = false }
-x509-cert = { version = "0.2.5", default-features = false, features = ["pem"] }
-der = { version = "0.7.9", default-features = false, features = ["alloc"] }
-ecdsa = { version = "0.16.9", default-features = false, features = ["digest", "alloc", "der"] }
-p384 = { version = "0.13.0", default-features = false, features = ["ecdsa-core", "ecdsa"] }
-hex = "0.4.3"
-
-[build-dependencies]
-fluentbase-build = { workspace = true }
-
-[lib]
-crate-type = ["cdylib"]
-path = "lib.rs"
-
-[features]
-default = ["std", "fluentbase-sdk/debug-print"]
-std = [
-    "fluentbase-sdk/std"
-]
-
