--- examples/nitro-verifier/Makefile
@@ -1,3 +1,0 @@
-.PHONY: lib.wasm
-lib.wasm:
-	cargo build
